# src/main/resources/esignsante-ws-api-doc.yaml
openapi: 3.0.0
...
servers:
-  description: SwaggerHub API Auto Mocking
   url: https://esignsante.asipsante.fr{basePath}/
   variables:
      port:
         enum:
         - '8080'
         default: '8080'
      basePath:
         default: /
info:
   description: "API du composant eSignSante. <br>Ce composant dit de \"signature\" mutualise et homogénéise la mise en oeuvre\
      \ des besoins autour de la signature. <br>Il permet de signer des documents ainsi que\
      \ de vérifier la validité d'une signature ou d'un certificat.    <br>"
   version: 2.3.0.0
   title: eSignSante
   contact:
      email: esignsante@asipsante.fr
...
paths:
...
   /signatures/xmldsig:
      post:
         tags:
         - xmldsig
         summary: Signature d'un document au format XMLdsig
         operationId: signatureXMLdsig
         description: |
            L'opération permet au client de signer un document au format XMLDsig-core-1.
         requestBody:
            content:
               multipart/form-data:
                  schema:
                     $ref: '#/components/schemas/SignDocument'
            required: true
         responses:
            '200':
               description: Requête en succès.
               content:
                  application/json:
                     schema:
                        type: object
                        $ref: '#/components/schemas/ESignSanteSignatureReport'
                        example:
                           erreurs: []
                           docSigne: PD94bWwgdmVyc2lvbj1cIjEuMFwiIGVuY29kaW5nPVwiVVRGLThcIj8+PGRvYz5bLi4uXTxkczpTaWduYXR1cmU+PC9kczpTaWduYXR1cmU+PC9kb2M+
            '400':
               description: Requête mal formée.
            '401':
               description: Une authentification est nécessaire pour accéder à la ressource (secret incorrect).
            '404':
               description: Ressource introuvable (identifiant de configuration qui n'existe pas par exemple).
            '500':
               description: Erreur interne du serveur.
            '501':
               description: L'opération n'est pas supportée par le serveur (soumission d'un document qui n'est pas au format
                  XML pour une signature enveloppée par exemple).
            '503':
               description: Le service n'est pas disponible.
   /signatures/xmldsigwithproof:
      post:
         tags:
         - xmldsig
         summary: Signature d'un document au format XMLdsig avec preuve
         operationId: signatureXMLdsigWithProof
         description: "L'opération permet au client de signer un document au format XMLDsig-core-1 ainsi que de générer une\
            \ preuve de vérification de signature. <br>Le client peut demander la vérification des règles applicables suivantes:<br>\
            \ &nbsp;&nbsp;&nbsp;validité de la signature du certificat;<br>&nbsp;&nbsp;&nbsp;présence de la balise \"SigningTime\" dans la signature;<br>   \
            \ &nbsp;&nbsp;&nbsp;présence du certificat utilisé dans la signature; <br>&nbsp;&nbsp;&nbsp;le certificat de signature n'est pas expiré; <br>&nbsp;&nbsp;&nbsp;l'usage\
            \ de la clé du certificat correspond à un usage de signature électronique et de non répudiation;<br>&nbsp;&nbsp;&nbsp;le certificat\
            \ de signature n'est pas révoqué; <br>&nbsp;&nbsp;&nbsp;présence d'une signature qui n'est pas vide; <br>&nbsp;&nbsp;&nbsp;l'Autorité de Certification\
            \ est reconnue;<br>&nbsp;&nbsp;&nbsp;le contenu de la signature est valide;<br><br>  Le client peut également demander l'extraction des\
            \ métadonnées suivantes: <br>&nbsp;&nbsp;&nbsp;le DN du certificat;<br>&nbsp;&nbsp;&nbsp;la date de signature;<br>&nbsp;&nbsp;&nbsp;le rapport du diagnostique;<br>  \
            \ &nbsp;&nbsp;&nbsp;le document original non signé.<br>"
         requestBody:
            content:
               multipart/form-data:
                  schema:
                     $ref: '#/components/schemas/SignDocumentWithProof'
            required: true
         responses:
            '200':
               description: Requête en succès.
               content:
                  application/json:
                     schema:
                        type: object
                        $ref: '#/components/schemas/ESignSanteSignatureReportWithProof'
                        example:
                           Rapport signature valide:
                              erreurs: []
                              metaData:
                              -  typeMetadata: DN_CERTIFICAT
                                 message: CN=asipsign.flux.asipsante.fr, OU=318751275100020, O=ASIP-SANTE, ST=Paris
                                    (75), C=FR
                              -  typeMetadata: DOCUMENT_ORIGINAL_NON_SIGNE
                                 message: uY29kaW5nPVwiVVRGLThcIj8+PGRv
                              valide: true
                              docSigne: PD94bWwgdmVyc2lvbj1cIjEuMFwiIGVuY29kaW5nPVwiVVRGLThcIj8+PGRvYz5bLi4uXTxkczpTaWduYXR1cmU+PC9kczpTaWduYXR1cmU+PC9kb2M+
                              preuve: PHhuczpBU0lQUHJvb2YgeG1sbnM6ZHM9XCJodHRwOi8vd3d3LnczLm9yZy8yMDAwLzA5L3htbGRzaWcjXCIgeG1sbnM6eG5zPVwiaHR0cDovL2VzYW50ZS5nb3V2LmZyL0FTSVAvUHJvb2ZcIiB4bWw6aWQ9XCJBU0lQX1Byb29mXCI
                           Rapport signature invalide:
                              erreurs:
                              -  codeErreur: ERSIGN02
                                 message: Le certificat n'a pas été émis par l'autorité de certification requise.
                              -  codeErreur: ERSIGN05
                                 message: Le certificat est expiré.
                              metaData: []
                              valide: false
                              docSigne: PD94bWwgdmVyc2lvbj1cIjEuMFwiIGVuY29kaW5nPVwiVVRGLThcIj8+PGRvYz5bLi4uXTxkczpTaWduYXR1cmU+PC9kczpTaWduYXR1cmU+PC9kb2M+
                              preuve: PHhuczpBU0lQUHJvb2YgeG1sbnM6ZHM9XCJodHRwOi8vd3d3LnczLm9yZy8yMDAwLzA5L3htbGRzaWcjXCIgeG1sbnM6eG5zPVwiaHR0cDovL2VzYW50ZS5nb3V2LmZyL0FTSVAvUHJvb2ZcIiB4bWw6aWQ9XCJBU0lQX1Byb29mXCI
            '400':
               description: Requête mal formée.
            '401':
               description: Une authentification est nécessaire pour accéder à la ressource (secret incorrect).
            '404':
               description: Ressource introuvable (identifiant de configuration qui n'existe pas par exemple).
            '500':
               description: Erreur interne du serveur.
            '501':
               description: L'opération n'est pas supportée par le serveur (soumission d'un document qui n'est pas au format
                  XML pour une signature enveloppée par exemple).
            '503':
               description: Le service n'est pas disponible.
   /signatures/xadesbaselineb:
      post:
         tags:
         - xades
         summary: Signature d'un document au format Xades Baseline B.
         operationId: signatureXades
         description: |
            L'opération permet au client de signer un document au format XADES Baseline B.
         requestBody:
            content:
               multipart/form-data:
                  schema:
                     $ref: '#/components/schemas/SignDocument'
            required: true
         responses:
            '200':
               description: Requête en succès.
               content:
                  application/json:
                     schema:
                        type: object
                        $ref: '#/components/schemas/ESignSanteSignatureReport'
                        example:
                           erreurs: []
                           docSigne: PD94bWwgdmVyc2lvbj1cIjEuMFwiIGVuY29kaW5nPVwiVVRGLThcIj8+PGRvYz5bLi4uXTxkczpTaWduYXR1cmU+PC9kczpTaWduYXR1cmU+PC9kb2M+
            '400':
               description: Requête mal formée.
            '401':
               description: Une authentification est nécessaire pour accéder à la ressource (secret incorrect).
            '404':
               description: Ressource introuvable (identifiant de configuration qui n'existe pas par exemple).
            '500':
               description: Erreur interne du serveur.
            '501':
               description: L'opération n'est pas supportée par le serveur (soumission d'un document qui n'est pas au format
                  XML pour une signature enveloppée par exemple).
            '503':
               description: Le service n'est pas disponible.
...
components:
   schemas:
...
      SignDocument:
         type: object
         required:
         - idSignConf
         - file
         properties:
            secret:
               description: Secret
               type: string
               pattern: ^[a-zA-Z_0-9\-]*$
            idSignConf:
               description: Identifiant de configuration à sélectionner parmi la liste des configurations disponibles pour
                  la signature (appel de l'opération "/configurations").
               type: integer
               format: int64
            file:
               description: Document à signer.
               type: string
               format: binary
      SignDocumentWithProof:
         type: object
         required:
         - idSignConf
         - file
         - idVerifSignConf
         - requestId
         - proofTag
         - applicantId
         properties:
            secret:
               description: Secret
               type: string
               pattern: ^[a-zA-Z_0-9\-]*$
            idSignConf:
               description: Identifiant de configuration à sélectionner parmi la liste des configurations disponibles pour
                  la signature (appel de l'opération "/configurations").
               type: integer
               format: int64
            file:
               description: Document à signer.
               type: string
               format: binary
            idVerifSignConf:
               description: Identifiant de configuration à sélectionner parmi la liste des configurations disponibles pour
                  la vérification de signature (appel de l'opération "/configurations").
               type: integer
               format: int64
            requestId:
               description: Identifiant de la demande pour renseigner l'élément RequestId de la preuve.
               type: string
               pattern: ^[a-zA-Z_0-9\-]*$
            proofTag:
               description: Tag utilisé pour renseigner l'élément Tag de la preuve.
               type: string
               pattern: ^[a-zA-Z_0-9\-]*$
            applicantId:
               description: Identifiant du demandeur utilisé pour renseigner le champ applicantId de la preuve.
               type: string
               pattern: ^[a-zA-Z_0-9\-]*$
...
      ESignSanteSignatureReport:
         type: object
         required:
         - erreurs
         - docSigne
         properties:
            erreurs:
               type: array
               items:
                  $ref: '#/components/schemas/Erreur'
            docSigne:
               description: Le document signé encodé en base 64.
               type: string
      ESignSanteSignatureReportWithProof:
         type: object
         required:
         - erreurs
         - metaData
         - valide
         - docSigne
         - preuve
         properties:
            erreurs:
               type: array
               items:
                  $ref: '#/components/schemas/Erreur'
            metaData:
               type: array
               items:
                  $ref: '#/components/schemas/Metadata'
            valide:
               type: boolean
            docSigne:
               description: Le document signé encodé en base 64.
               type: string
            preuve:
               description: La preuve de vérification encodée en base 64.
               type: string
...
      Erreur:
         required:
         - codeErreur
         - message
         properties:
            codeErreur:
               type: string
            message:
               type: string
      Metadata:
         description: Données complémentaires et paramétrables retournées par le service.
         required:
         - typeMetadata
         - message
         properties:
            typeMetadata:
               type: string
            message:
               type: string